+++ src/components/credit-editor.js
@@ -3,28 +3,23 @@ import imgflo from 'imgflo-url'
 
 import TextareaAutosize from './textarea-autosize'
 
-import DropdownWrap from './dropdown-wrap'
-import Arrow from 'rebass/dist/Arrow'
 import Avatar from 'rebass/dist/Avatar'
 
 
 export default function CreditEditor (props, context) {
   const {name, label, url, avatar, onChange, onlyUrl, path} = props
 
-  return el(DropdownWrap
-  , { buttonText: (name || label)
-    , menuIcon: el(Arrow, {direction: 'down'})
-    , menuKids: el('div'
-      , { style:
-          { padding: '1rem 1rem 0 1rem' }
-        }
-      , renderAvatar(avatar, context.imgfloConfig)
-      , (onlyUrl
-        ? renderBasedOnUrl(url, onChange, path)
-        : renderFields(name, label, url, avatar, onChange, path)
-        )
-      )
+  return el('div'
+  , { style:
+      { width: 360
+      , padding: '1rem 1rem 0 1rem'
+      }
     }
+  , renderAvatar(avatar, context.imgfloConfig)
+  , (onlyUrl
+    ? renderBasedOnUrl(url, onChange, path)
+    : renderFields(name, label, url, avatar, onChange, path)
+    )
   )
 }
 CreditEditor.contextTypes = {imgfloConfig: React.PropTypes.object}
@@ -50,21 +45,22 @@ function renderAvatar (avatar, imgfloConfig) {
 
 function renderFields (name, label, url, avatar, onChange, path) {
   return (
-    [ renderTextField('name', 'Name', name, onChange, path.concat(['name']))
-    , renderTextField('url', 'Link', url, onChange, path.concat(['url']))
+    [ renderTextField('name', 'Name', name, onChange, path.concat(['name']), true)
+    , renderTextField('url', 'Link', url, onChange, path.concat(['url']), false)
     ]
   )
 }
 
 function renderBasedOnUrl (value, onChange, path) {
-  return renderTextField('url', 'Link', value, onChange, path)
+  return renderTextField('url', 'Link', value, onChange, path, true)
 }
 
-function renderTextField (key, label, value, onChange, path) {
+function renderTextField (key, label, value, onChange, path, defaultFocus = false) {
   return el(TextareaAutosize
   , { className: `AttributionEditor-${key}`
     , label
     , defaultValue: value
+    , defaultFocus
     , key: key
     , name: key
     , multiLine: true